feat: type hierarchy
First pass at writing a type hierarchy option. The good news is that I haven't
broken the call hierarchy with this update. The bad news is that I don't run an
LSP that offers type hierarchy, so I cannot test this new code.

Until I can test it, this commit will sit here in this branch