refactor: Added an enum for direction
I also took the opportunity to move around the location of node, which didn't
make sense sitting in a folder of tree on it's own: the project is not
complicated enough to merit this sort of nesting!

I renamed telescope-ui to just ui. The old name was too long and didn't
meaningfully add to the semantics of what it was doing since _all_ UI is being
rendered through Telescope; that is literally the point of this extension

Finally I did a bit of documentation clean-up