refactor: add an enum for mode
Mirroring the structure of the enum we wrote for direction. I did discover that
we can type enums directly onto a table, as commented out in the code. This is
the standard way of doing enums but it doesn't (easily) allow me to attach
methods to instances of the enum (which I think makes the code cleaner). So I'm
sticking with my slightly convoluted home-grown solution for the time being

I took the opportunity of renaming 'hierarchy-type' to 'mode' in this change as
it's a shorter name that scans a bit better. I had wanted to call it 'type' but
that was no good as it both conflicted with the Lua LS annotations and was
confusing with the nomenclature for the type hierarchy that is to be implemented