fix: tidy up
With the correct Lua runtime configured in Lua-LS we can remove the deprecation
ignore. Tidy up the Lua annotation syntax in a couple of other places too

Adding folke/LazyDev as a dependency of my LSP set-up has also finally allowed
me to type Pickers and other objects from the Telescope library

Finally I was passing an argument to one of the Telescope functions that didn't
need it. Now removed