feat: cache entries
Small thing to the end user but required a reasonable extensive re-write. Now
when a function is "found" by the LSP, that find is remembered in a cache and
all other instances of that same function in the rendered tree will be shown
with the same remembered state

This involved splitting the memory of which functions call / are called by which
other functions into a hash map of locations to other functions in the cache.
Then the nodes have a link into the unique function cache for their location,
which means when one is found, they all are