fix: remove duplicates and nil results
Some LSPs (pyright) respond with duplicate entries in fromRanges which caused
me to render duplicated nodes on the treeview. I have put in some very basic
logic to deal with these dupes, which assumes they are always consecutive in the
reply list.

I also dealt with the issue that if a server responds with nil then that should
be treated as there are no entries for that call hierarchy