feat: goto definition
Telescope shows the place in code that the call is happening. It is sometimes
useful to be able to go to the definition of the function being shown instead of
where it gets called. I have added a new action and keymap to do just this.

Sadly I couldn't call into any Telescope function to do this as the Telescope
function that runs the edit command expects to work only with a Telescope entry
and read the filename, row and col from there. We need to read these values from
the location of the cache entry for this functionality to work. I have therefore
copied and ever so slightly amended the code from the Telescope plugin to fit my
purposes