docs: make settings clearer
I had an issue as a user (not unreasonably) copied the Install section code,
which then errored on the mappings section as telescope-hierarchy.actions was
not imported in scope. In fact telescope-hierarchy.actions could not be imported
in scope at that point as Lazy will not have put telescope-hierarchy in the
runtime path, so these instructions were fatal!

I have changed it round to make it clearer that the default copy and paste
instructions are to have no special settings at all.

On the mappings, I have for now taken them out. They should work but I need to
figure out the dance alluded to above which is at which point will the plugin
actually be loaded and available to be referenced? It will be somewhere in the
config function to be sure but that is complicated by telescope managing the
loading of the plugin and marshalling the application of it's opts. I neither
have the time nor energy to battle that dragon today