fix: callback even when LSP errors
It looks like we need to make the callback even when the LSP errors as those
callbacks have counting mechanisms in them that will fall over if not all of the
calls resolve

This was a little painful. We first needed to catch the right error and then in
that specific case, find the non-loaded file, load it, and then try again.

Hopefully will fix #9, but needs to be tested by a user that is suffering the
problem. I have been able to re-create the error locally so I am more hopeful
this is a viable fix but will await users who were suffering with the issue on
real projects first