feat: lexicographically sort incoming and outcoming calls
Currently calls are sorted in the order that the LSP tells us about
them. This can make it harder to find calls in a long list.

This PR makes sure that when children are added to a Node, that they are
then stored in a sorted order. In the comparison function that we use
with `table.sort` we make sure that anything in the current file is
sorted to the top of the list by representing it just as a number. For
references in any other file we concatenate the file name (normalised
from the cwd) with the line number, so that we first sort by file path,
then file name and finally by line number.

Storing the children sorted over sorting the children on each render is
preferred as we only find child nodes once but we potentially render
many times over

Closes #22, resolves #14